ignore e representacao de valores

+++ sells/serializers.py
@@ -6,7 +6,7 @@ class SellListSerializer(serializers.Serializer):
     date = serializers.CharField(read_only=True)
     products = serializers.CharField(read_only=True)
     client = serializers.CharField(read_only=True)
-    amount = serializers.IntegerField(read_only=True)
+    amount = serializers.DecimalField(max_digits=7, decimal_places=3,read_only=True)
     unit = serializers.CharField(read_only=True)
     price = serializers.IntegerField(read_only=True)
     products_storage = serializers.CharField(read_only=True)
